+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to the Aptos TypeScript SDK will be captured in this file. T
 - Populate `coinType` for `getAccountCoinAmount` if only `faMetadataAddress` is provided.
 - [`Fix`] `getModuleEventsByEventType` will also account for EventHandle events.
 - [`Hot Fix`] change regex to find object address when using `createObjectAndPublishPackage` move function
+- Add support for federated keyless accounts as defined in AIP-96
 
 # 1.27.1 (2024-08-23)